Optimize --table-re
The previous method would fetch all the data in the mysqldump and then
throw away the tables that did not match the --table-re. This change
excludes the tables from the mysqldump in order to avoid the extra work.
In my production database with 5000+ tables mysqldiff was taking about 5
minutes comparing tables outside the regex. It now tables about 2
seconds with this change.