Adding auto increment column fails with error 1075
Issue-30. Two statements are used to add an AUTO_INCREMENT primary key
column, and this causes a 1075 error: `Incorrect table definition; there
can be only one auto column and it must be defined as a key`

This revision adds checks for auto increment fields.  If a PRIMARY or
UNIQUE key has been defined for an auto increment field, it is appended
to the field definition instead of being created as a separate alter
statement.