support quoted fields
This supports quoted fields - but does require fields to be quoted in backticks - I think that is now a safe assumption. The objective is to support spaces in column names.
The other line processing regexes might need to be updated as they can now encounter backticks but this seemed to be sufficient for my database dumps.